QMake pro file for static lib building
As MinGW requires a lot of hassle with MSYS2, this pro file helps the build on the correct MinGW version for Qt developers.
The MSYS2 steps are still required for the configure step to generate the correct platform.hpp file, it is possible to use the delivered platform.hpp and adjust it manually.